addet HomeScreen

+++ app/src/main/java/com/elias/weatherapp/data/model/WeatherData.kt
@@ -2,6 +2,6 @@ package com.elias.weatherapp.data.model
 
 data class WeatherData(
     val temperature: Double,
-    val humidity: Double,
+    val humidity: Int,
     val windSpeed: Double
 )
